Removed unused interfaces
