+++ src/embedded/embeddedSessionManager.ts
@@ -1,5 +1,8 @@
 import { v4 as uuidv4 } from 'uuid';
-import { IEmbeddedSession, IEmbeddedImpression } from '../../src/types';
+import {
+  IEmbeddedSession,
+  IEmbeddedImpression
+} from '../../src/events/embedded/types';
 import { IEmbeddedImpressionData } from './types';
 import { trackEmbeddedSession } from '../events/embedded/events';
