+++ src/events/in-app/types.ts
@@ -1,8 +1,3 @@
-import {
-  IEmbeddedImpression,
-  IEmbeddedMessageElements,
-  IEmbeddedMessageMetadata
-} from '../../../src/types';
 export interface InAppTrackRequestParams {
   eventName: string;
   id?: string;
@@ -25,30 +20,4 @@ export interface InAppEventRequestParams {
   };
   inboxSessionId?: string;
   createdAt?: number;
-}
-export interface IEmbeddedMessage {
-  email?: string;
-  userId?: string;
-  messageId: string;
-  metadata?: IEmbeddedMessageMetadata;
-  elements?: IEmbeddedMessageElements;
-  payload?: Array<any>;
-  deviceInfo: {
-    appPackageName: string; // customer-defined name
-  };
-}
-
-export interface ISession {
-  start?: number;
-  end?: number;
-  id: string;
-}
-
-export interface IEventEmbeddedSession {
-  session: ISession;
-  placementId?: string;
-  impressions?: Array<IEmbeddedImpression>;
-  deviceInfo: {
-    appPackageName: string; // customer-defined name
-  };
 }
